feat: created getAll Clubs with minPrice,count and types of Court

+++ API-PI-Clubes/Application/DTOs/CourtDTO.cs
@@ -8,7 +8,7 @@ namespace API_PI_Clubes.Application.DTOs
         public TypeEnum Type { get; set; }
         public SurfaceEnum Surface { get; set; }
         public bool IsCovered { get; set; }
-        public double PricePerHour { get; set; }
+        public decimal PricePerHour { get; set; }
         public string Description { get; set; }
         public Guid ClubId { get; set; }
     }
@@ -19,7 +19,7 @@ namespace API_PI_Clubes.Application.DTOs
         public TypeEnum Type { get; set; }
         public SurfaceEnum Surface { get; set; }
         public bool IsCovered { get; set; }
-        public double PricePerHour { get; set; }
+        public decimal PricePerHour { get; set; }
         public string Description { get; set; }
     }
 
@@ -30,7 +30,7 @@ namespace API_PI_Clubes.Application.DTOs
         public TypeEnum Type { get; set; }
         public SurfaceEnum Surface { get; set; }
         public bool IsCovered { get; set; }
-        public double PricePerHour { get; set; }
+        public decimal PricePerHour { get; set; }
         public bool IsActive { get; set; }
         public string Description { get; set; }
     }
